backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pymongo==4.6.1
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.2.1
httpx==0.25.1
beautifulsoup4==4.12.2
python-multipart==0.0.6
aiohttp==3.9.3 